employee-sentiment-assistant/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
openai==1.30.1
httpx==0.27.0
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1
reportlab==4.2.0
pandas==2.2.2
aiofiles==23.2.1
bcrypt==4.0.1
